Some cleaning, added undulator exit fringe.
